newsboard/requirements.txt
django~=3.0
djangorestframework
django-cors-headers
psycopg2
gunicorn
honcho
schedule
